improve clear and init playground scripts

+++ _playground/clear-playground.sh
@@ -2,13 +2,15 @@
 cd "$(dirname "$0")"
 cd vscode-app-folder
 
-find . ! -name '.gitignore' -type f -exec rm -f {} +
+find . -not -name '.gitignore' -delete
+echo "All files deleted in vscode-app-folder"
 
-# Empty @worksapce cache folder if not empty already
-if [ -z "$(ls -A ~/.aws/amazonq/cache/cache)" ]; then
-    echo "Cache folder is empty"
-else
-    echo "Cache folder is not empty"
-    # remove all files in cache folder
-    rm ~/.aws/amazonq/cache/cache/*
+if [ "$1" == "--hard" ]; then
+    # Empty @worksapce cache folder if not empty already
+    if [ -z "$(ls -A ~/.aws/amazonq/cache/cache)" ]; then
+        echo "Amazo Q Cache folder is already empty"
+    else
+        echo "Amazo Q Cache folder is not empty"
+        rm ~/.aws/amazonq/cache/cache/*
+    fi
 fi
